Added Bootstrap carousel. CSS changes to work around issues with ".item'.
Fonts not needed for some reason (used for glyphicons), but I added them just in-case they are needed.

(*, *:before, *:after) -> I am not sure why added these in the first place. Removed as it does not affect functionality. I suspect it may have to do with consistency and compatibility accross different web browsers. Removed it because 'font-family' interfered with Glyphicons for the carousel. I could set a font-family specifically for 'glyphicon' in the CSS, but I deleted '*' instead because I don't think I needed it any way. Deleting that CSS will remove the Glyphicon font-family conflict.