Update portfolio. Fix masonry bug.
Added a much delayed update to the Change Log for v2.0.1 for Floating Point Calculator. Fixed a masonry bug where if you refresh the page, the masonry does not load and the page appears broken. The fix was simple, just move the masonry function outside of the .ready() function. Moved the most recent education and course history to the top. Changed the style of the masonry items only for the Computer Science courses.